requirements.txt
# api
uvicorn==0.22.0
fastapi==0.96.0

# database orm
SQLAlchemy==2.0.15
SQLAlchemy-Utils==0.41.1
psycopg2-binary==2.9.6

# schema models
pydantic==1.10.8

# .env
python-dotenv==0.19.2

# tests
pytest==7.3.1